setting up entitiies,controllers,repos,and services

+++ group assign co/roadco/src/main/java/groupassign/roadco/model/repositories/ClientRepo.java
@@ -2,10 +2,9 @@ package groupassign.roadco.model.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
 import groupassign.roadco.model.entities.Client;
 
 @Repository
-public interface ClientRepo extends CrudRepository<Client,Long>{
+public interface ClientRepo extends CrudRepository<Client,Integer>{
     
 }
